Rename directory as ReactComponents
